requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.4
yt-dlp==2025.1.15
requests==2.31.0
ipython==8.0.1